# api/openapi.yaml
openapi: 3.0.3
info:
  title: Symposium API
  version: 1.0.0
  description: |
    Read a Symposium community record, follow it as it changes, and publish through the gate.

    The Symposium API speaks in the specification's model: Artifacts, Objects, properties,
    relationships, Members and addresses (`spec/symposium_specification.md`, profiled by
    `tools/CANONICAL.md`). It is served by the Symposium Data server beside its file-level
    `/v1` API, and reads the community record held there, so every application sees the one
    record under the server's own access rules.

    **The record is read-only here.** Artifacts are immutable. A correction is a new Artifact
    that supersedes the old one, published through `POST /{community}/submissions` and decided
    by the gate. The API offers no update and no delete of a recorded Artifact.

    **Canonical URLs.** Every resource carries a `url`: an absolute URL on the data server that
    names it whichever host asks. An address maps to a URL by a fixed rule (see
    `GET /{community}/resolve` and `api/DESIGN.md`).

    **Freshness.** Every read response carries `position`: the record's cursor and the
    `created` of the newest Artifact the response reflects.

    **Paging.** Every listing whose size grows with the record takes `cursor` and `limit`
    (1 to 1000, default 100) and answers `next`, the cursor of the following page, or `null`
    on the last page. Each listing states its order.

...

...

...

...

...
  license:
    name: MIT
...
servers:
...
security:
  - apiKey: []
paths:
  /openapi.yaml:
    get:
      operationId: getOpenApiYaml
      tags: [contract]
      summary: This OpenAPI document, as YAML
      description: |
        The contract this server implements, exactly as committed in `api/openapi.yaml`. It
        answers with no credential, so any client can read it.

        Roles: non-member, member, admin; anonymous always.
      x-roles: [non-member, member, admin]
      x-anonymous: always
      security:
        - apiKey: []
        - {}
      responses:
        '200':
          description: The OpenAPI document.
          content:
            application/yaml:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /openapi.json:
    get:
      operationId: getOpenApiJson
      tags: [contract]
      summary: This OpenAPI document, as JSON
      description: |
        The same document as `getOpenApiYaml`, converted to JSON. It answers with no credential.

        Roles: non-member, member, admin; anonymous always.
      x-roles: [non-member, member, admin]
      x-anonymous: always
      security:
        - apiKey: []
        - {}
      responses:
        '200':
          description: The OpenAPI document.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /{community}/members:
    get:
      operationId: listMembers
      tags: [members]
      summary: The community's Members
      description: |
        Every handle on the roster, with whether it is registered and how many Artifacts it
        published by type. Ordered by handle ascending. This is `/symposium roster list`.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: One page of Members.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/members/{handle}:
    get:
      operationId: getMember
      tags: [members]
      summary: One Member
      description: |
        One Member. Its address is `@<handle>`. A Member exposes its name and nothing within it.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/handle'
      responses:
        '200':
          description: The Member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberResource'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/members/{handle}/artifacts:
    get:
      operationId: listMemberArtifacts
      tags: [members]
      summary: The Artifacts a Member published
      description: |
        Every Artifact whose `published_by` is this Member, ordered by `created` ascending.

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/handle'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: One page of Artifact summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactSummaryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/members/{handle}/messages:
    get:
      operationId: listMemberMessages
      tags: [members]
      summary: The Messages addressed to a Member
      description: |
        Every Message in the record whose `recipients` names this Member, ordered by `created`
        ascending. Messages are Artifacts of the record, so anyone who reads the record reads
        them. The gate's reply to a rejected submission is a different thing: it sits in the
        submission's `reply` field (`getSubmission`).

        Roles: non-member, member, admin; anonymous when the community's `record` collection is public.
      x-roles: [non-member, member, admin]
      x-anonymous: when the community's `record` collection is public
      security:
        - apiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/community'
        - $ref: '#/components/parameters/handle'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/order'
      responses:
        '200':
          description: One page of Message summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactSummaryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /{community}/me:
    get:
      operationId: getMe
      tags: [members]
      summary: Who this API key acts as
      description: |
        The key's username, its role, its scope and its expiry, and the Member it publishes as
        where it holds `member` or `admin`.

        Roles: non-member, member, admin.
      x-roles: [non-member, member, admin]
      x-bounded: one key
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/community'
      responses:
        '200':
          description: The key's identity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List every API key, with its value
      description: |
        Every API key, each with its username, its role, its scope, its creation, expiry and
        revocation, its last use, and the key itself, decrypted for this listing, so the admin
        can hand a lost key back. A revoked key's value is erased at revocation and answers
        `null`. Ordered by creation ascending. This is `/symposium list-api-keys`.

        Roles: admin, through the server admin's Ed25519 token only.
      x-roles: [admin]
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
        - name: community
          in: query
          description: Only keys scoped to this community.
          schema:
            type: string
            pattern: '^[A-Za-z0-9_]{1,20}$'
        - name: username
          in: query
          description: Only keys for this username.
          schema:
            type: string
        - name: include_revoked
          in: query
          description: Include revoked keys.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: One page of API keys.
          headers:
            Cache-Control:
              description: Always `no-store`, because the page holds secrets.
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create an API key
      description: |
        Create a key for a username with one role and a scope, and answer it once with its
        value. A `member` key must name a handle registered on its community's roster, and
        publishes as that Member. An `admin` key names the server's admin handle and is bound
        to the admin's current Ed25519 key. A key scoped to the server holds `admin`. This is
        `/symposium gen-api-key <username> <role>`.

        Roles: admin, through the server admin's Ed25519 token only.
      x-roles: [admin]
      x-bounded: one key
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: The new key, its value included.
          headers:
            Location:
              description: The key's canonical URL.
              schema:
                type: string
                format: uri
            Cache-Control:
              description: Always `no-store`, because the body holds a secret.
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /admin/api-keys/{key_id}:
    get:
      operationId: getApiKey
      tags: [api-keys]
      summary: One API key, with its value
      description: |
        One key, as `listApiKeys` shows it.

        Roles: admin, through the server admin's Ed25519 token only.
      x-roles: [admin]
      x-bounded: one key
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/keyId'
      responses:
        '200':
          description: The key.
          headers:
            Cache-Control:
              description: Always `no-store`, because the body holds a secret.
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke an API key
      description: |
        Revoke a key. It stops authenticating at once, its encrypted value is erased, and its
        row stays for the record of who held it. Revoking a revoked key answers it unchanged.
        This is `/symposium revoke-api-key <key id>`.

        Roles: admin, through the server admin's Ed25519 token only.
      x-roles: [admin]
      x-bounded: one key
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/keyId'
      responses:
        '200':
          description: The revoked key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sak_<43 base64url characters>
      description: |
        An API key the admin issued, sent as `Authorization: Bearer sak_…`. The `sak_` prefix
        tells it apart from the data server's Ed25519 tokens (JWTs) and read keys (`sdr_…`),
        which share the same header. The server matches a key by its SHA-256 digest; it
        decrypts nothing to authenticate.
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The server admin's Ed25519 token from `POST /v1/admin/token`, the one the skill already
        holds. The API-key operations accept it alone, so an API key, whatever its scope,
        reads and makes no other key.
  parameters:
    community:
      name: community
      in: path
      required: true
      description: The community's name, matched ignoring case.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_]{1,20}$'
...
    handle:
      name: handle
      in: path
      required: true
      description: The Member's handle, its `name` without the `@`.
      schema:
        type: string
        pattern: '^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$'
...
    keyId:
      name: key_id
      in: path
      required: true
      description: The API key's id.
      schema:
        type: string
        format: uuid
    cursor:
      name: cursor
      in: query
      description: The `next` of the previous page. Absent for the first page.
      schema:
        type: string
        maxLength: 256
        pattern: '^[A-Za-z0-9_-]+$'
    limit:
      name: limit
      in: query
      description: The most items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    order:
      name: order
      in: query
      description: '`created` for oldest first, `-created` for newest first.'
      schema:
        type: string
        enum: [created, -created]
        default: created
...
  responses:
    BadRequest:
      description: A parameter or the body is malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: The credential is absent, unknown, expired or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The credential is valid and its roles or scope do not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The community or the resource is unknown.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Unprocessable:
      description: The body is well formed and its values break a rule, which `detail` names.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: |
        The one error body. `detail` is a sentence for a person, as the data server's `/v1`
        writes it; `code` is stable for a program.
      required: [detail, code]
      properties:
        detail:
          type: string
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - too_large
            - validation_failed
            - unprocessable
            - not_operational
            - too_many_streams
        findings:
          type: array
          description: The validator's findings, on `validation_failed` only.
          items:
            $ref: '#/components/schemas/Finding'
    CommunityName:
      type: string
      pattern: ^[A-Za-z0-9_]{1,20}$
    Handle:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$
    Name:
      type: string
      description: A name in the specification's sense. It holds no `.` or `#` and starts with no `@`.
      pattern: ^[^@.#][^.#]*$
    Address:
      type: string
      description: |
        An address (specification §1.8): `@` and dot-separated segments naming a Member, an
        Artifact, an Object or a property, then optionally `#` and a schema reference.
      pattern: ^@[^.#]+
      example: '@agent_lyra_data_overexpression_screen_v1.values#csv.row=BST2&col=Normalized infection'
    Cursor:
      type: string
      maxLength: 256
      pattern: ^[A-Za-z0-9_-]+$
      description: |
        An opaque position in an ordered listing, base64url. It stays valid for the life of the
        record.
    Position:
      type: object
      description: The state of the record a read response reflects.
      required: [cursor, seq, created, as_of]
      properties:
        cursor:
          $ref: '#/components/schemas/Cursor'
        seq:
          type: integer
          minimum: 0
          description: The `record` collection's sequence number at the newest Artifact reflected.
        created:
          type: string
          nullable: true
          format: date-time
          description: The `created` of the newest Artifact reflected; null for an empty record.
        as_of:
          type: string
          format: date-time
          description: When the server read the record for this response.
...
    Page:
      type: object
      description: The fields every paged listing carries.
      required: [next, position]
      properties:
        next:
          description: The cursor of the following page; null on the last page.
          type: string
          nullable: true
          maxLength: 256
          pattern: '^[A-Za-z0-9_-]+$'
        position:
          $ref: '#/components/schemas/Position'
...
    ArtifactSummary:
      type: object
      description: One Artifact's header fields and counts, as a listing carries it.
      required: [name, address, url, type, published_by, created]
      properties:
        name:
          $ref: '#/components/schemas/Name'
        address:
          $ref: '#/components/schemas/Address'
        url:
          type: string
          format: uri
        type:
          type: string
        title:
          type: string
        published_by:
          $ref: '#/components/schemas/Handle'
        created:
          type: string
          format: date-time
        groundable:
          type: boolean
          description: False for a non-groundable type or a header saying so.
        supersedes:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        superseded_by:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        counts:
          type: object
          properties:
            objects:
              type: integer
            grounds:
              type: integer
            tests:
              type: integer
              description: Grounds carrying a `criterion`.
            findings:
              type: integer
    ArtifactSummaryPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ArtifactSummary'
...
    MemberResource:
      type: object
      required: [handle, address, url, registered, counts, position]
      properties:
        handle:
          $ref: '#/components/schemas/Handle'
        address:
          $ref: '#/components/schemas/Address'
        url:
          type: string
          format: uri
        registered:
          type: boolean
          description: True once the Member has joined from its invite.
        counts:
          type: object
          properties:
            artifacts:
              type: integer
            by_type:
              type: object
              additionalProperties:
                type: integer
        artifacts_url:
          type: string
          format: uri
        messages_url:
          type: string
          format: uri
        position:
          $ref: '#/components/schemas/Position'
    MemberPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/MemberResource'
    KeyRole:
      type: string
      description: |
        The key's one role. `member` holds everything `non-member` holds, and `admin` holds
        everything `member` holds.
      enum: [non-member, member, admin]
    KeyScope:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [community, server]
        community:
          $ref: '#/components/schemas/CommunityName'
    Me:
      type: object
      required: [username, role, scope, key_id, position]
      properties:
        username:
          type: string
        role:
          $ref: '#/components/schemas/KeyRole'
        scope:
          $ref: '#/components/schemas/KeyScope'
        key_id:
          type: string
          format: uuid
        expires:
          type: string
          nullable: true
          format: date-time
        member_url:
          type: string
          format: uri
          description: The Member this key publishes as, where it holds `member`.
        position:
          $ref: '#/components/schemas/Position'
...
    ApiKeyCreate:
      type: object
      required: [username, role, scope]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64
          description: |
            For `member`, a handle registered on the scope's roster. For a server-scoped
            `admin` key, the server's admin handle. For `non-member`, a label naming the
            application, which matches no handle on the roster.
        role:
          $ref: '#/components/schemas/KeyRole'
        scope:
          $ref: '#/components/schemas/KeyScope'
        expires_days:
          type: integer
          minimum: 1
          maximum: 366
          description: Absent for a key that lasts until revoked.
        label:
          type: string
          maxLength: 200
    ApiKey:
      type: object
      required: [id, url, username, role, scope, key, created, created_by, expires, revoked, last_used, uses]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        username:
          type: string
        role:
          $ref: '#/components/schemas/KeyRole'
        scope:
          $ref: '#/components/schemas/KeyScope'
        key:
          type: string
          nullable: true
          pattern: ^sak_[A-Za-z0-9_-]{43}$
          description: The key. Null once revoked.
        label:
          type: string
        created:
          type: string
          format: date-time
        created_by:
          type: string
        expires:
          type: string
          nullable: true
          format: date-time
        revoked:
          type: string
          nullable: true
          format: date-time
        revoked_by:
          type: string
          nullable: true
        last_used:
          type: string
          nullable: true
          format: date-time
        uses:
          type: integer
    ApiKeyPage:
      type: object
      description: |
        A page of API keys. Keys belong to the server, outside any record, so the page carries
        no `position`.
      required: [items, next]
      properties:
        next:
          description: The cursor of the following page; null on the last page.
          type: string
          nullable: true
          maxLength: 256
          pattern: '^[A-Za-z0-9_-]+$'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApiKey'