# data-server/service/symposium_api/generated/routers/api_keys.py
# generated by fastapi-codegen:
#   filename:  api/openapi.yaml

"""Generated from api/openapi.yaml by codegen/generate.py. Never edit by hand."""

from __future__ import annotations

from fastapi import APIRouter, Depends
from fastapi.responses import StreamingResponse

from symposium_api.authz import Caller, authorize
from symposium_api.provider import service_provider

from ..dependencies import *
from ..service import Service
from ..sse import SSE_HEADERS, sse_frames

router = APIRouter(tags=["api-keys"])


@router.get(
    "/admin/api-keys",
    response_model=ApiKeyPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listApiKeys",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
    },
)
def list_api_keys(
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    community: Optional[str] = None,
    username: Optional[str] = None,
    include_revoked: Optional[bool] = True,
    caller: Caller = Depends(authorize("listApiKeys")),
    service: Service = Depends(service_provider),
) -> ApiKeyPage:
    """List every API key, with its value"""
    return service.list_api_keys(
        caller,
        cursor=cursor,
        limit=limit,
        community=community,
        username=username,
        include_revoked=include_revoked,
    )


@router.post(
    "/admin/api-keys",
    response_model=ApiKey,
    response_model_exclude_unset=True,
    status_code=201,
    operation_id="createApiKey",
    responses={
        "201": {
            "model": ApiKey,
        },
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
        "422": {
            "model": Error,
        },
    },
)
def create_api_key(
    body: ApiKeyCreate,
    caller: Caller = Depends(authorize("createApiKey")),
    service: Service = Depends(service_provider),
) -> ApiKey:
    """Create an API key"""
    return service.create_api_key(caller, body=body)


@router.get(
    "/admin/api-keys/{key_id}",
    response_model=ApiKey,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getApiKey",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_api_key(
    key_id: UUID,
    caller: Caller = Depends(authorize("getApiKey")),
    service: Service = Depends(service_provider),
) -> ApiKey:
    """One API key, with its value"""
    return service.get_api_key(caller, key_id=key_id)


@router.delete(
    "/admin/api-keys/{key_id}",
    response_model=ApiKey,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="revokeApiKey",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def revoke_api_key(
    key_id: UUID,
    caller: Caller = Depends(authorize("revokeApiKey")),
    service: Service = Depends(service_provider),
) -> ApiKey:
    """Revoke an API key"""
    return service.revoke_api_key(caller, key_id=key_id)