# data-server/service/symposium_api/generated/routers/contract.py
# generated by fastapi-codegen:
#   filename:  api/openapi.yaml

"""Generated from api/openapi.yaml by codegen/generate.py. Never edit by hand."""

from __future__ import annotations

from fastapi import APIRouter, Depends
from fastapi.responses import StreamingResponse

from symposium_api.authz import Caller, authorize
from symposium_api.provider import service_provider

from ..dependencies import *
from ..service import Service
from ..sse import SSE_HEADERS, sse_frames

router = APIRouter(tags=["contract"])


@router.get(
    "/openapi.json",
    response_model=OpenapiJsonGetResponse,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getOpenApiJson",
    responses={
        "401": {
            "model": Error,
        },
    },
)
def get_open_api_json(
    caller: Caller = Depends(authorize("getOpenApiJson")),
    service: Service = Depends(service_provider),
) -> OpenapiJsonGetResponse:
    """This OpenAPI document, as JSON"""
    return service.get_open_api_json(caller)


@router.get(
    "/openapi.yaml",
    response_model=str,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getOpenApiYaml",
    responses={
        "401": {
            "model": Error,
        },
    },
)
def get_open_api_yaml(
    caller: Caller = Depends(authorize("getOpenApiYaml")),
    service: Service = Depends(service_provider),
) -> str:
    """This OpenAPI document, as YAML"""
    return service.get_open_api_yaml(caller)