# data-server/service/symposium_api/generated/routers/members.py
# generated by fastapi-codegen:
#   filename:  api/openapi.yaml

"""Generated from api/openapi.yaml by codegen/generate.py. Never edit by hand."""

from __future__ import annotations

from fastapi import APIRouter, Depends
from fastapi.responses import StreamingResponse

from symposium_api.authz import Caller, authorize
from symposium_api.provider import service_provider

from ..dependencies import *
from ..service import Service
from ..sse import SSE_HEADERS, sse_frames

router = APIRouter(tags=["members"])


@router.get(
    "/{community}/me",
    response_model=Me,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getMe",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_me(
    community: str,
    caller: Caller = Depends(authorize("getMe")),
    service: Service = Depends(service_provider),
) -> Me:
    """Who this API key acts as"""
    return service.get_me(caller, community=community)


@router.get(
    "/{community}/members",
    response_model=MemberPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listMembers",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_members(
    community: str,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    caller: Caller = Depends(authorize("listMembers")),
    service: Service = Depends(service_provider),
) -> MemberPage:
    """The community's Members"""
    return service.list_members(caller, community=community, cursor=cursor, limit=limit)


@router.get(
    "/{community}/members/{handle}",
    response_model=MemberResource,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="getMember",
    responses={
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def get_member(
    community: str,
    handle: str = ...,
    caller: Caller = Depends(authorize("getMember")),
    service: Service = Depends(service_provider),
) -> MemberResource:
    """One Member"""
    return service.get_member(caller, community=community, handle=handle)


@router.get(
    "/{community}/members/{handle}/artifacts",
    response_model=ArtifactSummaryPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listMemberArtifacts",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_member_artifacts(
    community: str,
    handle: str = ...,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    order: Optional[Order] = "created",
    caller: Caller = Depends(authorize("listMemberArtifacts")),
    service: Service = Depends(service_provider),
) -> ArtifactSummaryPage:
    """The Artifacts a Member published"""
    return service.list_member_artifacts(
        caller,
        community=community,
        handle=handle,
        cursor=cursor,
        limit=limit,
        order=order,
    )


@router.get(
    "/{community}/members/{handle}/messages",
    response_model=ArtifactSummaryPage,
    response_model_exclude_unset=True,
    status_code=200,
    operation_id="listMemberMessages",
    responses={
        "400": {
            "model": Error,
        },
        "401": {
            "model": Error,
        },
        "403": {
            "model": Error,
        },
        "404": {
            "model": Error,
        },
    },
)
def list_member_messages(
    community: str,
    handle: str = ...,
    cursor: Optional[str] = None,
    limit: Optional[int] = 100,
    order: Optional[Order] = "created",
    caller: Caller = Depends(authorize("listMemberMessages")),
    service: Service = Depends(service_provider),
) -> ArtifactSummaryPage:
    """The Messages addressed to a Member"""
    return service.list_member_messages(
        caller,
        community=community,
        handle=handle,
        cursor=cursor,
        limit=limit,
        order=order,
    )